Updates for M1 library changes

+++ web/hello_debug/hello_debug.dart
@@ -12,18 +12,18 @@ DebugDrawManager _debugDrawManager;
 
 int _viewport;
 Camera _camera;
-int _lastTime;
+double _lastTime;
 bool _circleDrawn = false;
 
-void frame(int time) {
+void frame(double time) {
   if (_lastTime == null) {
     _lastTime = time;
     window.requestAnimationFrame(frame);
     return;
   }
-  int dt = time - _lastTime;
+  double dt = time - _lastTime;
   _lastTime = time;
-  double seconds = dt.toDouble() * 0.001;
+  double seconds = dt * 0.001;
   // Update the debug draw manager state
   _debugDrawManager.update(seconds);
   // Clear the color buffer
